support message variants
This commit adds basic support for message variants.  Variants
are placed in `variant/*.cmake` files.  Variants can include
different sections of the message.  There is rudimentary support
for encoding differences.  Currently the only option supported
is whether the `|` and `$` syntax is considered to be syntactic
sugar that needs expanding into parentheses, or genuine symbols
to be communicated.